almost done adding controllers for now playing
