+++ app/src/main/java/me/mikecasper/musicvoice/nowplaying/events/StartQueueFragmentEvent.java
@@ -1,0 +1,4 @@
+package me.mikecasper.musicvoice.nowplaying.events;
+
+public class StartQueueFragmentEvent {
+}
